toplink essential depends on persistence API 1.0b, which upgrades the entire dependency used in v3 to 1.0b.
This causes a problem at runtime, since many modules have Class-Path entry to 1.0, yet it's not found. We need more metadata in jar so that this can be resolved to 1.0b at runtime.

But until I do that, this bug is blocking others, and this fix will let us get going.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@17621 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5