Renamed the PE bundle to the web bundle. Added the 'glassfish' bundle, which contains web+ejb+acc+...
Relevant distribution bundle building scripts are refactored for better reuse,
and maven-antrun-extended-plugin is bumped up to 1.14 during this process.

With this change, the version of maven-antrun-extended-plugin is now given at
distribution/pom.xml, instead of sprinkling the same value all over various child POMs.
Technically speaking, version in maven plugin can be always omitted to indicate
the latest, but that will leave it up to individual Maven to decide when to check for
plugin updates. Our set up in distribution/pom.xml allows us to force downloading
newer version of the plugin.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@17731 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5