Switching to Woodstox.
The actual dependency is specified in HK2, and relevant POMs
are modified to avoid specifying explicit dependencies to SJSXP
(to avoid having both at the same time.)

This is still bit error prone in the sense that if anyone
adds SJSXP dependency accidentally, we'd end up getting
both jars in the distribution, which would be a waste.
Ideally, we should detect this sort of situations.

All the unit tests and quicklook runs are confirmed.
Note that you won't pick up Woodstox until you pick up
a newer version of HK2.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@18780 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5