DomainXml: It uses the instanceName (i.e. the "name" in <server> element) that is sent in via args to Main if GFLauncher is the caller then the instance-name is guranteed to be set correctly. If GFLauncher is not used then this arg may or may not be set in which case "server" is used -- It used to always use "server"
V3Environment now has:

getDomainName() -- this is the domain name parsed from the value in domain.xml -- if GFLauncher is used and if it *is* in domain.xml.
Otherwise we assume we are starting a DAS instance and use the name of the root dir

getInstanceName -- see above

getProps()
returns the props that were parsed out of asenv[.bat|conf], it also has com.sun.aas.instanceRoot set.
Use this in preference to System Properties
It is a static final



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@18841 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5