ActionReport is an interface ActionReporter is an abstract class that implements ActionReport
ActionReporter has essential methods not found in the interface -- e.g. setContentType and writeReport

That effectively forces any class that wants to implement ActionReport to instead extend ActionReporter
In which case -- why have an interface at all?

Note that AdminCommands are really stuck because the caller operates with an ActionReporter but Admin Commands only get an ActionReport.

I added some methods from ActionReporter to the interface.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@19453 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5