Issue 4614
Start-domain waits for DAS by repeatedly calling the "version" command until it gets a response.
The code gets the admin port from domain.xml
The adminport is sent in as a parameter instead of an option, so the version command ignores it and always runs version on port 8080.
The fix was just to put "--port" before the port number

Before the fix, if your DAS was not listening on 8080 -- then DAS would start, but you'd get a 90 second timeout while it waits for a response it will never get.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@19609 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5