Have org.apache.catalina.valves.ValveBase implement the Tomcat Valve interface (in addition to the GlassFish Valve interface).
This avoids ClassCastException (in GlassFish) when loading a class
that extends Tomcat's ValveBase (which implements Tomcat's Valve
interface) and casting it to the Tomcat Valve interface.

In org.apache.catalina.valves.ValveBase, the additional methods from
the Tomcat Valve are implemented as no-ops, but any of the GlassFish
container valves that may be used as "basic valves" (positioned at the
end of their respective pipeline), and will be implemented as "hybrid"
valves, will override these methods as appropriate.

Tests run: QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@21212 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5