Good programming practice, use Java interfaces only to define types, not constant literals [EJ-2 19].
This is our public API. Striving hard to get it right first time.
Interfaces should not be used to define constants literals (at least
in product's public interface).

It's not clear to me why this class is a "public" API of GlassFish
since it defines some string literals, but that's a separate story.

Bumping up the glassfish-api version.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@21371 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5