This bug was lying in wait. JPA woke it up.
java.io.File.listFiles() has the vexing property of doing two different things in the 2 kinds of negative cases:

(1) if the directory has no files in it -- it returns an empty array
(2) if the directory is not really a directory -- a NPE is thrown

Embedded located this bug.  I deployed a simple war that has no class files.  Thus it has no WEB-INF/lib directory.
Nevertheless File.listFiles() was called on the bogus directory anyways.

Fix: make sure the file is a bona fide directory before using it.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@21559 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5