Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=3710 ("Restore Virtual Server's Root Context Behavior from GlassFish v1"):
Create a request facade such that if the request was received at the
root context, and the root context is mapped to a default-web-module,
the default-web-module mapping is masked from application code to
which the request facade is passed. For example, the request facade's
getContextPath() method would return "/", rather than the context root of
the default-web-module, in this case.

Tests run: QL, Servlet TCK

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22204 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5