GlassFish: 5760
Lo and Behold!

I have added 9 more tests(!) to the QL suite. It adds
less than a second to the overall running time. Now
onwards, the number of tests would be 24. 

Passed all the 24 tests.

Newly added tests are:
- creation/deletion/listing of http-listeners.
- creation/ping/deletion/listing of JDBC-connection-pools.

More tests are needed, but don't worry, adding a new asadmin
test is very easy now (IMO, give it a try).



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22267 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5