Fix for 5605 - Security: Using directory deployment on an application in a folder using non-latin characters exposes the user's hard drive.
Peter has checked in changes for the NB plug-in to use & for separator among query parameters, rather than ? which it now uses only for introducing the query string.  This check-in changes the server side to look only for & as the separator.  (It used to look for both & and ? as a temporary transition until all clients stopped using ? as a separator.)

Tests: QL and manual tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22318 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5