Fix for 5810 - Deployment fails if dir for app in ${domain-dir}/applications already exists
On Windows especially a JAR file might remain locked and cannot be deleted, which also means the containing directory tree cannot be deleted (as normally happens during undeployment).  This fix to the deployment code path tolerates a lingering directory and warns about it but allows the deployment to continue.  Locked files that remain behind are overwritten in place (which Windows does allow).

Reviewed: Jerome

Tests: QL on both Windows and Mac; manual tests on Windows to ensure a locked file is overwritten and the new contents is used correctly

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22352 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5