Integrating HK2 version 0.3.20 to fix Issue #5681, #5934: First time when we attempted to fix 5681, we did so by setting changing any new bundle's start level to be higher than that of frameworks. This has an undesirable side effect that those bundle don't get activated even when we call bundle.start() until the framework's start level reaches bundle's start level. We rely on web-glue bundle to be activated so that it could register a URLStreamHandlerService to handle jndi scheme. Since it was not happening, we started seeing issue #5934.
As noted in bug #5861, we had made that change temporarily, we wanted to switch
to bundle.start(START_TRANSIENT) as soon as Felix supports this. Since it was
causing the above regression, I have backed out that change. Instead, we now 
maintain explicit list of bundles that we start in HK2 OSGi management agent.
We use a boolean field called isTransientlyActive in every module object to 
keep this information. When ModulesRegistry shuts down, we stop all those
bundles for which isTransientlyActive is true. When ever HK2 starts a bundle,
we mark that bundle with isTransientlyActive = true. 


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22465 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5