GlassFish: 5722.
Added the support for --keytooloptions cn=common-name-of-the-distinguished-name-for-server-cert.

See the bug report for details.
Tested thoroughly, locally. Made sure that proper certificate messages
are shown by the Firefox browser on Mac OS X. 

QL does not test this and I'll soon add a test for this.

Now, we should have a server certificate that gets created correctly,
for "localhost" rather than some arbitrary server.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@22609 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5