GlassFish: 6215
We were throwing away all the exception messages. For example,
if you create an http-listener with port not in range [0, 65535],
you won't know if that's not the correct value. This is based
on the datatype annotation given on @Configured interfaces.

In the processes, corrected some other classes as well, to
make sure we are not throwing away the knowledge from underlying
invocations.

QL runs. Tested manually. Need to increase the test coverage, but
that needs time ...



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@23059 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5