IT 5518
add package dependencies.

Reviewed: Snjezana

Verified the changes by installing one package and the dependent packages are also installed.
To get the server up and running the following two packages are needed:    
  * glassfish-jpa
  * glassfish-management

Tested this by creating and starting the domain with asadmin.
Verifed that the domain is running by executing "asadmin list-commands" and http://<host-name>:8080.

To get admin console up and running, need to install the following packages:
  * glassfish-gui
  * glassfish-web-management 

Verified the login page with admin console url.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@23157 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5