This is quite a Saxy checkin...
1) SaxParserHandlerFactory.java --> Embedded and presumably ACC broke because of a check for whether or no installRoot 
is set.  installRoot is now set  earlier in the startup which caused this problem.  
Fix:  see if there is a directory "lib/dtds" that exists under the installRoot.  If not -- then fetch DTDs from
the internet.  I.e. don't just assume that if installRoot is not null  that it is valid.


2) pom.xml --> Add embedded module back in

3) DeployWebAppTest is a Embedded Unit Test.  Removed the @Ignore annotation.  Embedded Unit Tests  are now once again running.

4) QL build.xml
While trying to debug (1) above I set up a debug build of JDK6 from Sun.  It apparently spits out a "hotspot.log" file everytime the JVM starts.  
QL tried to delete that file (*.log) which failed and thus QL instantly and always failed.
I removed the command to delete *.log with Ming's blessing.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@23463 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5