Porting rev #23437 from v3-prelude-release branch to trunk. Details: 0. The big change: all versions have been updated to use our new version number: 3.0-SNAPSHOT. When final rel will be made, we shall use 3.0. 1. glassfish-api extends glassfish-parent. This simplifies glassfish-api/pom.xml. e.g., We no longer have to separately configure the plugins, maintain dependency versions, etc. It now uses same version as parent. If you are wondering, how other projects will depend on glassfish-api, then answer is they will depend on promoted builds of glassfish-api. It is upto us to decide promotion frequency.
2. build (maven-glassfish-extensions and maven-glassfish-plugin) module now
extends glassfish-parent. This simplifies their poms. They all use same
version as their parent. Since v3 build depends on these extensions, we
have to introduce a profile to avoid "chicken and egg" problem. More over,
we had to remove maven-glassfish-plugin from top level pom to pluginManagement section.
The top level pom had earlier defined an extension for and maven-glassfish-extensions. This introduces
yet another circular dependency that maven can't handle. So, we have moved the extension definition
to every pom that relies on these extensions.

3. Removed unused properties like pe.version, dist.version, security-api.version.

4. grizzly versions are now maintained in one place only.

5. HK2 versions are maintained in one place only.

6. Removed redundant properties like amx-api.version, glassfish.version, jdbc-ra.version. Their usages
have been replaced with ${project.version}.

7. Fixed scm tags.

8. We are temporarily using 0.1 version of org.glassfish.docs:basedocs instead of relying on 0.2-SNAPSHOT.
When 0.2 is released, we need to change nucleus/pom.xml.

9. Other minor build improvements like moving dependencies to top level, using externalizing OSGi metadata
out of pom.xml to osgi.bundle file in glassfish-api module, etc.

10. Added javax.servlet:3.0 as a dependnecy in web-glue/pom.xml otherwise maven sometimes uses servlet-api:2.5
in compilation classpath ahead of servlet 3.0 apis and that results in compilation failures.

With all these changes, I am able to do a clean build and run QL.
I am still not able to use maven-release-plugin to prepare a release as it fails tto locate
maven-glassfish-extension. I am checking with maven community about this. But, we can use 
this plugin to do branching, verify that there are no SNAPSHOT dependencies, etc. So, I think
it is a good improvement from where we were.

I have not tried to build packager module.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@23482 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5