Porting back changes from commit 23476 to branches/v3_prelude_release. Here are comments from that commit JPADeployer code iterates through all the BundleDescriptors and gather all referenced pus in them and create emf for each of them. When ejbs are inside war, we have two BundleDescriptors in the Application object(Web and EJB). This results in two distinct emfs being created and used for same pu. This can result in data corruption
This checkin fixes the issue

Since for prelude we can never have pus being overridden (no ears), it is safe to initialize only one emf per pu name and put it in Application object
instead of referring BundleDescriptor.

The changes in JPADeployer.java ensures that only one pu of given name gets enlisted for creation.
The changes in JPAApplication puts the emf in Application object instead of referring BundleDescriptor. 

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@23521 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5