Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=7158
This needed integration of HK2 version #0.3.39.
While exporting packages, we use a property called project.osgi.version to
specify the package version. This property is populated during build by
a HK2 maven plugin called osgiversion-maven-plugin. That plugin is configured
to drop qualifier while converting the maven version to osgi version as
a result of which we can now use builds from trunk with promoted builds
and vice-versa.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@24657 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5