This was really a bug-fix in search of a bug!
FindBugs complained that BASE64Decoder/Encoder was "shadowing" sun.misc classes of the same name which was the intention 
and the whole point.

I tried fixing this properly by copying over BASE64 code.  I wasted 5 hours, and discovered that it just was not worth it to
spend any more time (money) on something that works perfectly well.

My solution was to change the name of the classes.  We still have the option of moving the implementation whenever we choose in the 
future.  This will make FindBugs happy.





git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@25744 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5