Introduced support for specifying a jdbc connection pool property : "Prefer-Validate-Over-Recreate" that takes either a true or false as its value.
By setting this property, the pool resizing is optimized. When this property is set to true, the pool resizer checks whether an idle connection (candidate for refreshing) from the steady part of the pool's free list is still valid and only destroys and recreate an idle connection if it has been reported as being invalid. The resource timestamp is set to a new value after the connection has been tested for validity.

A counter : validConnectionsCounter is incremented as long as it is less than the steady pool size and the connection is valid. When the connection is invalid, the resource is removed directly.

Without this flag, the physical connection that is idle in the pool will be destroyed and recreated in all cases without checking for its validity. This happens even if this connection should stay alive due to the steady pool size setting. This flag when set to true, takes advantage of the fact that the underlying ManagedConnectionFactory supports connection validation (i.e, instanceof ValidatingManagedConnectionFactory).

This is a per pool property and is valid for all JDBC and Connector connection pools. This improves optimization by validating the idle connections from the steady pool size part and only then decide to destroy/recreate them if they are invalid.

Reviewed by :
Jagadish Ramu

Tests Run:
Unit tests while setting property "Prefer-Validate-Over-Recreate" to true. Acquire connections, sleep, let resizer do the wor
k and acquire connections once more. Check for the physical connection IDs and compare them to be the same.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@25862 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5