Fix for 7985 (package-appclient missing)
Because v3 does not ship with ant, package-appclient is now in Java instead of an ant script.  Note that because Java's API for jar and zip entries does not offer access to setting permissions, the appclient script included in the JAR which package-appclient creates does not have execute permission when it is extracted.  Users will need to set the permission explicitly after extracting the contents of the generated file.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@26620 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5