Fix for 7656 (misleading error if directory deployed app contains unexpanded submodules)
Added logic makes sure the submodule archive is not an InputJarArchive.

Tests: QL, deployment devtests

Jane reviewed the pom change (needed because I added LocalStrings.properties to the resources branch; the pom was explicitly including only the java branch - and there are no resources files there to include.  The pom change restores the default plug-in behavior.)

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@26638 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5