I put in a quick work-around fix for a catastrophic P1 bug (I don't know if it has been reported).
The bug was creating a deadlock due to Logging issues and a JDK bug.  This causes the appserver to appear frozen.  Forever.
Mac users with heavy-duty iron see this bug and are not able to get the latest JDK where it is (alledgedly) fixed.
The deadlock can still EASILY happen by other code -- but it will not happen because of LogDomains code anymore!



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@26788 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5