Fix for issue: 8206
The default panels in open installer that handle JDK Selection and detection
are updated to fix this issue.

The detection process takes more time in slower machine and due to lack of 
progress notification, it would appear as if the installer is hanging. This
was presenting a bad user experience. The fix has moved the detection logic
to "view", so that cursor shape can be controlled. The detection logic has
been removed from "model".


 


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27333 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5