Fix for internal CR 6839668 - Fail CTS samples with java.lang.ClassCastException
Part of building gf-client.jar extracts the manifest from that JAR, updates it to add some entries, then updates the manifest in the JAR so it contains the additions.  The ant zip task to update the manifest in the JAR would not always work, for two reasons.  

1. ZIP files store entry timestamps to two-second resolution.  It is possible that the manifest entry in the JAR would not predate the revised one by at least two seconds, in which case the ZIP ant task would not recognize the entry as obsolete and would not replace it.  

2. The ant manifest task, used to update the Class-Path in the manifest, seems not to update the file revision timestamp of the on-disk manifest file.  So the zip ant task would never recognize the revised manifest as newer than the corresponding entry in the JAR file.

This fix introduces a three-second sleep before updating the extracted manifest (to make sure the zip task will judge it to be newer than the entry in the JAR) and also uses the ant touch task to force an update to the manifest file's revision timestamp.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27346 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5