https://glassfish.dev.java.net/issues/show_bug.cgi?id=8387
stop-domain:  
Before: if there is no domain running, it is an error.
After: If the final state of the Domain is not-running -- then we return SUCCESS no 
matter *how* we got to that state.

start-domain:
Before: If domain is already running, return an error
After: Only return n error if the final state is no running domain

This behavior is essential for getting Platform Services to properly manage AS Lifecycle.

SQE Tests have been changed to respect this.
QL tests, I discovered, completely ignore the return values so it had no effect on QL tests.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27534 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5