https://glassfish.dev.java.net/issues/show_bug.cgi?id=8381
create-domain can now tell the difference now between a port that is already in use and a port that the user is not allowed  to use.  
A classic example of the latter is a non-root user on UNIX trying to set a http listener port to 80.

There is no change in the behavior but different error messages are produced for the different scenarios.
Before:  the error message was "port in use" for both scenarios.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27567 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5