Fix for 8483 - app client sniffer claims non-.jar modules containing Main-Class
The sniffer formerly always would claim to handle any module with a manifest containing a Main-Class entry.  This specifically includes the Hudson .war file .  With this change, the sniffer will claim such modules only if the file type ends with .jar (or .jar/ or _jar/ for directory-deployed apps or submodules which are expanded).  The sniffer will always claim a module if it contains an application-client.xml or sun-application-client.xml file.  

Reviewed: Hong

Tests: QL, EJB and deployment devtests; Hudson deployment

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27738 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5