Result of svn merge -r17383:27767 from branches/asadmin. Reason: The trunk development is easier when we have multiple developers. Doing simultaneous branch/trunk development is a pain with maven unless we use maven release plugin which is also a pain ;).
This is the first and probably the only merge from asadmin branch to trunk. Has the
implementation of new cli protocol. 

All QL passed. Learned that osgi.bundle is an important file, although I think
the default should have been to export all the packages encountered in a given
svn module. Alas ...



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@27777 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5