Adding a new feature of the jdbc/connector connection pools : Pooling.
The new attribute pooling when set to false will disable connection pooling for the specified pool. 

By default, this attribute's value will be true. The existing system property 'com.sun.enterprise.connectors.SwitchoffACCConnectionPooling', that turns off connection pooling in ACC and the new attribute pooling do not affect each other.

Associate-with-thread functionality is invalid when pooling is disabled and hence an exception will be thrown if it is set to true. An exception will be thrown if Flush connection pool is done when pooling is disabled.

A warning will be logged when the following attributes are active as they are useful in pooled environment only :

    * connection-validation
    * validate-atmost-once-period
    * match-connections
    * max-connection-usage
    * idle-timeout 

Reviewed by : Jagadish Ramu

Tests run : QL, jdbc devtests, connector devtests, pooling devtests.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@28303 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5