Simple change.  The response to the condition changed:
CONDITION:
The btrace jar does not exist AND monitoring is enabled

BEFORE:
FATAL error -- we print out a message about the missing jar and where it is supposed to be.

AFTER:
We still check for the condition but just silently deal with it.
We do NOT give the JVM a "-javaagent" arg -- since the JVM will choke and not run.
We silently do this -- no error, no warning



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@29860 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5