Fixed issue 8940, 8604. Also ported fix of issue 4431.
In the previous delegation model, EarClassLoader was the parent of the 
module classloaders and it also delegated to module classloaders which 
led to the potential of classloading deadlock. With this set of the 
changes, the EarClassLoader will no longer delegate to the module 
classloaders which will prevent the deadlock from happening.

Ran QL, deployment dev, ejb dev tests (with known failures).


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@31356 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5