Partial fix for 9164
The appclient scripts needed to define additional properties so that security would work.  The policy file needed to specify the JARs in the GF modules directory (and its subdirs).  

With these changes in place, the ACC gets farther during launch, but the permissions failure now occurs in the GF ORB code as it tries to perform a privileged operation with non-priv'd app client code on the stack.  It'll need a doPrivileged block.

Tests: Q:, ejb and deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@31429 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5