Fix for 8726 and 9092
The ACC expects the generated JAR files to contain GlassFish-specific entries in the manifest, but was not careful about making sure that (1) the manifest existed and (2) that the entries were present; it assumed both were OK leading to NPEs in some cases when problems occurred during deployment and the generated JARs were incomplete or improperly formed.

These changes are much more defensive and more informative if the generated JARs are not to the ACC's liking.

Tests: QL:, ejb and deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@31556 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5