GlassFish:9498 GlassFish:9592
Optimize the code path that affects the "default" startup sequence.
In the default case, we first try to see if "changeit" is the
master password. This saves about 240 msec "every time" on my quad
core machine. Added some other changes that make the server not
verify the master password because start-domain has already done
it for it. In java -jar we continue to see the 240 sec penalty.

Code logic change discussed with Bill.

Removed some extraneous statements, localized the messages.

QL passes, 100%.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@31915 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5