Fix for 9946
These changes make the ClientArtifactManager more defensive.  It now throws exceptions when a deployer attempts to register an artifact at the same relative path as a previously-registered artifact or when a deployer tries to register an artifact for which the on-disk file does not exist or cannot be read.

Tests: QL:, deployment devtest

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@32204 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5