Making the current resource adapter deployment descriptor "sparse" so that it is possible to use only the descriptor to override what is provided as annotations.
* The current schema requires the deployer/developer to provide a 
lot of redundant information even when the deployer/developer only 
wants to override or augment the information provided by annotations.

To realize this ease of development feature we have to reduce number 
of required elements in the descriptor schema. The following
elements are now made optional:
- Under outbound-resourceadapterType:
    - connection-definition: If this is not specified, we require
the application server to use the connection definitions specified 
through @ConnectionDefinition. If no @ConnectionDefinition is specified,
this must be an error and end up in deployment failure.
    - transaction-support: If unspecified, the application server must 
treat this as NoTransaction. This is the default in the @Connector 
annotation as well
    - reauthentication-support: The earlier 1.5 spec talked about this 
being optional, but minOccurs was not set to zero.
- Under connectorType:
    - vendor-name: If unspecified, default "" as in @Connector
    - eis-type:  If unspecified, default "" as in @Connector
    - resourceadapter-version:  If unspecified, default "" as in @Connector

This would help a deployer to use a minimal(sparse) DD (a test is checked in 
to demonstrate this) to just override/augment annotation-specified metadata information. 

* A one-line describing the test is added in the existing tests



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@32604 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5