IT 10269 I also made some improvements while i had the hood up.
FlashlightProbeProvider is mainly a C style struct that contains 3 Strings.  All sorts of *other* code is constantly checking and verifying on
whether these strings actually have a value.  So I changed the strings to be final and the constructor now guarantees that each string will have some sort of
valid value.  Robustness starts at the foundations :-)  Lots of other code can be simplified now...

The way you get a "qname" from a PP is messy and duplicated.  I moved the code over to FlashlightUtils.makeName() which should be the one and only
place where names are calculated.

Now if you try to register a PP that has already been registered -- no problem-o.  It used to get you an Exception and you could not figure out whether or not
it is a "real" error...







git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33027 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5