IT 10264
If(the domain requires a Master Password to start && the only way to get it is to prompt the user && we are running as a service)
    print out a better error message 

Here is the new error message:

no.console=The Master Password is required to start the domain.  No console, no prompting possible.  You should either create the domain \
with --savemasterpassword=true or provide a password file with the --passwordfile option.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33334 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5