https://glassfish.dev.java.net/issues/show_bug.cgi?id=9934
JDK6 bug.  Discovered by trying to build with JDK7

Superclass method is trying to access a private variable of superclass through a reference to a subclass.  This was never allowed in Java.
The compiler in 6 must be confused by the Generics

Fix:  Simply upcast the subclass reference. (Suggested by Bill Shannon)



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33337 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5