Security permissions required to run JRuby inside GlassFish when SecurityManager is turned on.
The assumes Jruby is installed at install-root (glassfishv3/glassfish), which would be the case when JRuby is installed from the updatecenter. For the cases where JRuby is installed elsewhere, this grant needs to be replicated with the different JRuby location in the CodeBase.

Fixes, https://glassfish.dev.java.net/issues/show_bug.cgi?id=10271.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33440 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5