Fix the location of the GlassFish-provided endorsed JARs in the gf-client.jar Class-Path setting.
The ACC makes use of a masking class loader to make sure that the GlassFish-provided versions of classes are used, rather than their counterparts that are also provided in the Java runtime.  The ACC (and the package-appclient command) relies on the Class-Path setting from gf-client.jar being correct, but it did not correctly point to where the endorsed JARs actually reside - in modules/endorsed rather than in modules/ itself.  

These changes correct that.

pom.xml and fixup.xml changes reviewed by Jane - thanks


Tests: deployment and selected ejb devtests, QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33672 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5