Fix for 10732
The drive letter on Windows needs to be upper-case for security to work (!).  Depending on how the appclient script was invoked, the drive letter might be upper- or lower-case.  It must be upper-case.  This change forces the drive letter to upper case (WIndows only).


Tests: deployment and app client devtests; Sherry's help with the failing sqe tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33695 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5