Incremental fix for 9579
During deployment, annotation processing of an app client includes any JARs referenced from the app client JAR's manifest Class-Path.  In this issue, the app client refers to an EJB module (basically treating it as a library JAR).  The EJB module contained annotations that make sense in an EJB but not in an app client, so the EntityManagerReferenceHandler logged a severe message but continued on.  

What needs to happen to really resolve this is that the handler needs to know not only what the current processing context is (that is, what type of module is being examined - in this case an app client) but also whether it is currently processing that original JAR or a library JAR referenced  from that original JAR.  (Such references might be implicit with the JAR residing the EAR's library directory or explicit by appearance in the appc lient JAR's manifest Class-Path.) 

That sort of repair might not be feasible for the 3.0 release.  We're changing the severity of the error to warning (from severe) but not changing any other logic.  Later on we can revisit this.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@33698 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5