Added module-name related changes to the schema and updated connector-complete test to include module-name. Also fixed the container requirements around sparse DD
In EE 6, all Java EE modules have a name. (See Section 8.1.1 of the EE 6
specification describes this and also the rules for determining the
default module name). In connectors, its only use is to let the developer/deployer 
set the desired module name in the deployment descriptor. This has been added
to ra.xml schema for consistency reasons.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@34091 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5