Fix for 11241
For generated interfaces the protection domain was fetched but not stored, so that if/when used later it was null.  (This surfaced for me during tests of Java Web Start launches of app clients.)  The change simply stores the retrieved protection domain into the already-existing private field.

Tests:  ejb and deployment devtests; QL tests
Reviewed by: Ken S.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35104 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5