Further refactored osgi-javaee modules. Common container utilities have been moved to osgi-javaee-base. osgi-web-container and osgi-ejb-container provided specialized containers.
Added initial implementation of EJB/OSGi integration.
So, given a EJB OSGi bundle, we can recognize it and deploy it in underlying javaee container.

We need a fix in deployment/common/.../DeploymentContext.java to avoid a NPE.

Todo: register EJBs as services, allow wrapping of classic ejb jar as bundle.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35270 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5