Fix for 11294 (a Windows-only problem)
If the developer builds an EAR containing a submodule directory (not a submodule JAR) using the name that the server uses to expand submodules during deployment, then on Windows problems would occur because the server would try to expand the submodule from its current location into that very same location.  The logic for copying archives would try to overwrite an existing open file (since the same file was being read and written at the same time).  

This works fine on non-Windows systems, but it fails on Windows and the copy logic was not reporting any problem.  The net result was that each file that the server tried to copy was empty (zero length).  

These changes work around this problem by avoiding the expansion of the submodule if the submodule is packaged inside the EAR as a directory and if it has the name it would have if the server did expand it.

Reviewed by: Hong
Tests: deployment devtests, QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35555 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5