1. Switched to using official OSGi APIs artifacts as maven dependencies. 2. OSGi/HTTP service module now uses service tracker to track starting of glassfish. 3. Upgrading to HK2 1.0.1. It includes a cleaner StartupContext interface. StartupContext used to have a couple of methods called getRootDirectiry() and getUserDirectory(). Depending on the runtime, either both were explicitly set or both were unset. When both were not set, getRootDirectory() used to return a default value of System.getProperty("user.dir") and getUserDirectory() used to return null. More over, it was not clear what value should be returned by those calls. In some cases (e.g., in non-embedded case), getRootDirectory() returned path to glassfish/modules/, where as in embedded case, it returned path to glassfish/. Over all, it was very confusing. So, I decided to clean it up. It no longer has any of those methods. It has a properties object which can be set upfront to set the context. We expect individual bootstrap modules to store appropriate context information there for latter retrieval. Such information can include things like com.sun.aas.installRoot, com.sun.aas.instanceRoot, etc.
The second change in HK2 is the way HK2 OSGi adapter module is configured. Earlier, it used to be configured
via system properties. Now, we use OSGi config admin API to configure that bundle. When that bundle starts,
it does nothing until it receives a configuration update which it uses to create a StartupContext and Habitat.

4. Cleaned up ASMain* classes. Got rid of unnecessary abstraction like AbstractMain. glassfish.jar has now its own 
activator which is uses to configure HK2 bundle. It reads startup context information from bundle context (or system
properties) and then configures HK2. It is also responsible for starting any other bundles needed by glassfish. 
Such bundles include config admin, shell, file install, etc.

5. I have temporarily commented out jruby test in QL, as we have to integrate a new version of grizzly-jruby-module.jar.
We are waiting for RE to releaes that artifact from glassfish-scripting project after which we can integarte that version and
uncomment the test.

Tests performed: QL, embedded tests, QL on equinox


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35609 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5