Fix for 11071.
To support Java Web Start launches of app clients, the server traverses the app client's manifest Class-Path and extensions references (and their transitive closures) to make sure that the referenced JARs are accessible so they can be downloaded when needed.  The checking logic did not account for directory deployments in which the referring app client is in the generated directory (while the referenced JARs are still in the developer's directory).  These changes fix that.

In addition, the server now logs a single warning message (without a stack trace) for any JAR or extension that is referenced but not found where expected.  The server continues processing but the warning helps to alert the developer that one or more files he or she thought were present are not - at least not where specified.

Tests: deployment and ejb  devtests; QL; handful of manual Java Web Start tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35826 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5