Fix for 11649
In some cases (e.g., launching an app client from a program using the embedded API) the required ACCClassLoader would not have been instantiated but some code assumed it was already there.

This check-in fixes that and also improves a log message reminding a developer of a program using the embedded API to pass the URI to the client archive with the "jar:" scheme if the archive is in fact a JAR and with the "file:" scheme if the archive is a directory.

Tests: deployment and ejb  devtests; QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@35881 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5