Fix occasional unit test failures as reported by Byron.
The simplePermanentTransferDirTest incorrectly captured the lastModified value from the original directory before the file in the directory was created.  Usually, apparently, the directory creation and the file creation (and attendant directory time update) would happen during the same clock tick, so the lastModified on the directory would not actually change within the resolution of the clock.  But sometimes the clock tick would intervene and although the lastModified on the transferred directory would actually be correct the value would not match the (incorrect) captured value.

The change captures the original directory's lastModified value after the file is created in the directory, so the original directory's lastModified should always match the transferred directory's value.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@36581 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5