Fix recurring "null" output lines during JSR-88 execution.
Some of the recent changes in commit 36467 to admin/cli RemoteCommand.java cause the command output to be logged unless explicitly requested otherwise.  This change was not reflected in the RemoteDeploymentFacility, so JSR-88 (which uses RemoteCommand) began logging nulls in some cases.  

This change makes RemoteDeploymentFacility invoke the new executeAndReturnOutput method (which suppresses the logging).

Tests: deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@36607 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5