Fix for 11829
The value stored in the file-xfer-root property in payload parts is intended to be expressed as a file path (not a URI) using the file path syntax of the destination system.  Some SQE tests accidentally (but fortunately) specified paths for downloading client files that included braces { }.  The payload handling did not process the file-xfer-root correctly as a file path, assuming instead it was already a valid URI.  This led to problems in creating URIs for storing the transferred files.

This fix repairs that problem.

Tests: deployment devtests, with upload on and off; new unit test to exercise the target file path

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@36833 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5