Unify names going out of REST layers to camelCasedNames.
This is implemented on way out of REST layer by:  
1. converting xml-names from serverbeans to camelCasedNames 
2. Assuming that alias attribute from @Param are in camelCased form and corresponds to respective serverbean attribute names. 

On way into REST layer by:
1. convertring camelCasedNames to xml-names for server beans attributes
2. mapping alias to actual command param name

Modified CreateJdbcConnectionPool such that  alias for each @Param corresponds to respective serverbean attribute name. 

Tests run quicklook. There are AMX failures in test run that does not seem to be because of my changes.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@37137 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5