Greatly shortens the time over which an input stream is open to a file that is being added to the payload.
Previously, the PayloadImpl would open an input stream at the time a file is submitted for inclusion in the payload.  For a large payload, such as during remote instance sync operations, a relatively long time could pass before the stream was actually used to write the file's contents to the payload, at which point the stream was closed.  

These changes cause the PayloadImpl to open an input stream for an attached file immediately before the stream will be used to read the file's contents.  (PayloadImpl will continue to close the stream immediately after the file's contents have been added to the payload.)

Tests: QL, deployment devtests with and without --upload=true

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@38627 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5