Issue 12762 missing <monitoring-service/> in domain.xml causes all monitoring statistic to 0
When <monitoring-service /> *exists* in domain.xml, MiniXmlParser checks for monitoring-service element.  If monitoring-enabled is missing or "true", then GFLauncher will set this jvm option:
-javaagent:C:/fishfry/v3/distributions/glassfish/target/glassfishv3/glassfish/lib/monitor/btrace-agent.jar=unsafe=true,noServer=true

When there is *no* <monitoring-service />, MiniXmlParser does not find the monitoring-service element and the monitoringEnabled flag stays initialized to false.
And -javaagent is not set.

btrace handles probe transformation.  Without btrace, this probe is not created and fired, and listener is not called.
The fix is to initialize the monitoringEnabled flag to true in MiniXmlParser.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39000 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5