Additional fix to allow off-line builds to work.
The earlier fix for this worked when build was run from the appclient/client/acc-config directory but not from any higher directory - such as the top-level, for example.

This fix uses a full path to point to where the extracted DTDs reside.  (The earlier fix used a relative path that worked during the build from that directory only.)

pom change reviewed by Jane


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39076 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5