Fix for 12887
An earlier change in the ACC relaxed how aggressive the ACC would be in running clean-up logic.  This caused problems with JMS clients because MQ creates non-daemon threads as connections are opened from the client to queues/topics.  As non-daemon threads, they would prevent the VM from exiting unless the connections were closed or unless System.exit was invoked.  

This check-in restores the more aggressive clean-up in the ACC. We might be able to return to the less aggressive approach if the MQ implementation changes to use daemon threads.

Tests:  QL, deployment dev tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39583 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5