12640 Rename "nodeagents" dir to "nodes"
This change renames the "nodeagents" directory to "nodes". I have
run quicklook and the admin cli devtests and they passed. Since
this change is fairly wide-ranging it is possible I missed something.
If something breaks and you suspect it is because of this change
try creating a "nodeagents" symlink that points to "nodes". If that
works around the problem let me know and I'll investigate.

Warning!!! With this change GlassFish will not be able to run
on instances created with earlier 3.1 builds. You either need
to create new instances or rename the "nodeagents" directory to
"nodes" manually. This change handles renaming the directory on
v2 upgrades, but not from earlier 3.1 builds.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39660 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5