Additional fix for 12909
Deployments of osgi modules do not create generated directories.  The code which prepares the generated content zip for upload to instances failed if there were absolutely no generated files and directories.  With this change deployment will not create the zip file at all if there is nothing to zip, and it will not specify the --generatedcontent option to the instance deploy command either.  

Tests: QL, example OSGI module deployment to an instance

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39681 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5