Additional check-in for secure admin implementation
Refactored some logic from CreateServiceCommand that is now shared with CreateLocalInstanceCommand.  

CreateLocalInstanceCommand now passes more info to SecureAdminClientManager to construct the domain config if a domain.xml is available, given the parameters on the command line.  (If the user specifies a node or nodedir or server name the command now will try to build the domain config so it can see if secure client admin is enabled, etc.

Tests: QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@39736 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5