Issue #12913: Support fragments for WABs.
We now have a specialised version of Archive called WAB that adapts an OSGi host bundle 
and its associated fragments to a Web Application structure that's defined in the Servlet
Specification. e.g., all Bundle-ClassPath entries of type jar are presented as some jar in
WEB-INF/lib/; all Bundle-ClassPath entries of type directory (. excluding) are presented
as WEB-INF/classes. Bundle-ClassPath entry of type dot (.), is treated like a jar type entry.
The code now handles fragment Bundle-ClassPath entries as well and
it is able to correctly handle injection of Bundle-ClassPath entry from a host to fragment.
Example of the latter is: 
host has BCP: extra.jar, default.jar
fragment jar has an entry called extra.jar.

We now also are able to look up static resources in jars in Bundle-ClassPath. Static resources
in bundle-classpath jar must be packaged under META-INF/resources/ as per Servlet 3.0 spec.
To implement this, we have to configure repositories correctly in WebappClassLoader. See
issue 12990 for the reason.

We are also able to discover web-fragments.xml in jars in Bundle-ClassPath jars.

I have not paid attention to performance impact of doing all these, so we may have to look at
that in future if that becomes an issue. In particular, if a WAB is deployed using
asadmin deploy --type=osgi, we are not able to take advantage of the exploded dir created by
deployment backend, because that's unaware of the fragments. We should look at optimizing
our implementation when there is no fragment attached at least.

I have run WABs containing jsps, jsfs, jpa, ejbs and they all pass.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40146 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5