Fix for 13139
If the administrator did not specify an explicit installation directory at node creation time, then the secure admin bootstrapping would not provide the correct default installation directory.  (The installation directory is used for finding where the "nodes" directory should be, under which will appear the directory for the specific node in question.)

This check-in fixes that.

Tests: QL, deployment EE devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40258 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5