fix for issue 12961.
The problem here is the use of backslashes in the commands that run over SSH.  There are now additional commands that return the Unix style paths or backslashes replaced with forward slashes.  In addition, the code that runs the commands has been updated to change the backslashes to forward slashes in case the user manually entered a path.  

All QL tests passed on Unix.  Manual tests were run on windows.  These failures show up when communication is between Windows boxes.  Unix to Windows tests were working.  


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40335 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5