Changed the way hybrid apps get deployed. Earlier we have separate containers responsible for deployment of hybrid apps. We had a WebExtender which was monitoring bundle life cycles events and when it detects WABs, it used to hand it over to OSGiWebContainer to do the deployment. Similarly, EJBExtender was detecting EJB enabled bundle (manifest entry: Export-EJB) and doing necessary deployment into GlassFish EJB container. The issue here was that OSGiWebContainer and OSGiEJBContainer would both try to deploy if the same bundle has both a Web-ContextPath and Export-EJB headers. To avoid this double deployment issue, I am reorganising the runtime like this:
There is on container for all kinds of hybrid apps - it's the OSGiContainer in osgi-javaee-base module.
It gets activated by a new extender called JavaEEExtender. This container tracks OSGiDeployers which
are registered by osgi-ejb-container and osgi-web-container bundles. JavaEEExtender tracks bundles
and when bundles become ready, it calls handles() on each deployer. Whichever deployer is found first
is given the chance to deploy. To enforce an ordering (which is needed for use case explained below),
OSGiContainer uses service ranking to sort the deployers. 

Why we need a predefined order to decide which deployer should be used to deploy?
Imagine a WAB contaning some EJBs. So, it has both Web-ContextPath as well as Export-EJB
header. Given that Servlet spec has strict rules about structure of a web application,
we have to chose OSGiWebDeployer to deploy. So, currently, OSGiWebDeployer is registered
with highest service ranking and OSGiEJBDeployer is deployed with lowest.

The code correctly handles dynamically addition and removal of osgi dpeloyers.

With all these changes, I am able to run WABs, EJB enabled bundles, CDI enabled WABs, etc.

Also fixed a bug in WABClassLoader. We were not setting delegate=true while constructing the
WABClassLoader and the default in WebappClassLoader is false. This was causing a strange issue 
that Siva pointed out for his WAB.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40357 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5