Fix upgrade issue that Bobby encountered in cluster upgrade. Verified that this fixes following issue.
v2 cluster attribute heartbeatport is randomly generated over a much larger range of values than v3.1.
Thus, it is quite possible that a generated heartbeatport from v2 would cause a ValidationException when
set into v3.1 cluster attribute gmsMulticastport.  Fix is to catch ALL exceptions when setting 
a value from v2 to v3.1 just in case there is the same issue for other values being upgraded.

Both gmsMulticastport and gmsMulticastaddress are not allowed to be unset so if v2 value is outside of v3 valid range,
generate port or address approrpiately.

Tested these changes with a heartbeatport of 33000.  This value is outside of range allowed for v3.1 gmsMulticastport.
The checked in fix will generate a valid gmsMulticastport within the acceptable v3.1 range.

When this issue happened in past, ALL cluster modifications were not made since there was no try catch blocks
setting individual values. To be completely defensive and not allow one mismatch to break everything, the
fix is to catch the possible RuntimeException hk2 validation exception whenever setting v3.1 value 
with a v2 value.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40374 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5