Issue 12616 - fix problems with running admin devtests on Windows. The ClusterTest is modified so that it run validate-multicast before running the GMS get-health tests. If multicast is not available, those tests are skipped and a warning message is printed.
The step to delete the files from a previous run has been moved to a "clean" target which is no longer 
automatically run by init. This prevents a startup problem on Windows

The backup tests have been commented out of the normal run because the backup code sets incorrect file 
permissions on Windows which causes many test failures.  See issue 13225. Once that is fixed, the backup tests can be run again. 


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40390 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5