Fix for issue 13040
EJBContainerServicesImpl wasn't considering a super-class as a valid business interface for a no-interface
view of a session bean. Adding a method to EjbAbstractDescriptor to store/retrieve those supeclasses
for the bean and populating this list once in AbstractEjbHandler.

Tests run: PE and EE QL, EJB and CDI developer tests, and added a dev-test 
[cdi/javaee-component-resources/no-interface-ejb] to cover the scenarios listed
in the issue.

Thanks to Marina for reviewing the patch


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40563 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5