Update upgrade to Geenerate cluster property GMS_LISTENER_PORT-<cluster-name> just as v3.1 asadmin create-cluster would create it.
Cluster v2 to v3 Upgrade instructions must instruct users that each server element must have a system-property GMS_LISTENER_PORT-<cluster-name>
with a unique ununsed port for the machine it is going to be run on.  If each instance is going to run on own machine,
no need to set the server system proeprty.  However, each instance that is to run on same machine must set this value,
just as they have to when running multiple v3.1 instances on a single machine. 

Also no need to set this system property if v2 heartbeat-enabled = false and v3 gmsEnabled is set to false. 

This is a new requirement for v3.1 (due to grizzly transport).

*****

Code was literally copied from admin/config-api/Cluster.java.
Verified that it output the property during upgrade.r





git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40917 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5