A CDI portable extension to bridge OSGi services and Java EE platform - an initial version
A Java EE component (such as servlet, EJB etc) or a Bean in a hybrid application could 
include this portable extension and express a dependency on an OSGi
Service in the OSGi Service Registry, by qualifying the [@Inject based] injection point 
with the @OSGiService qualifier. The portable extension discovers all such 
injection points and then finds the OSGi service in the Service Registry, binds to it 
and enables the injection of the Service to that injection point.

Tests run: The advanced WAB in the JavaOne Hands On Lab [S313522] was modified to use this
extension



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@40959 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5