1. Making changes to simple-glassfish-api as per ASARCH review: Fixed javadocs. Add GlassFishException to method signature. Use var-args in Deployer and CommandRunner. Make CommandRunner.run return CommandResult. (Still need to discuss exit code issue) Make all public methods of GlassFishRuntime non-static except bootstrap(). Added a convenient bootstrap() method in GlassFishRuntime so that user does not have to pass a classloader. Removed destroy() from RuntimeBuilder - a GlassFishRuntime can do necessary cleanup in its shutdown().
We still need to clean up all the Constants and Options.

2. Implementation changes:
Changed implementation to accommodate the above api changes.
We now have a working Deployer and CommandRunner.
Moved ConfiguratorImpl to bootstrap module and made it a POJO as there are issues looking it up
inside GlassFishImpl.
Moved CommandExecuterImpl and DeployerImpl to admin/cli module, as they have dependency on Parser
class from that module. We don't want kernel to depend on admin/cli.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@41785 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5