Issue #14123: We were not correctly mapping an OSGi bundle with non-default Bundle-ClassPath to an archive. e.g., when a bundle has a Bundle-ClassPath: bin/, the mapped archive contained entries like bin/com/acme/MyBean.class although this class belongs to com.acme package. As a result, EJB sniffer was not able to find any EJBs. To fix the issue, we have now introduced an abstraction called EJBBundle similar to WAB. I refactored most of the code out of WAB to an abstract called OSGiJavaEEArchive and WAB as well as EJBBundle extend from that. As a result of this change, now even EJB bundles support fragments just like WAB.
I also found minor issues in our effective bundle classpath computation logic. e.g., we were
not handling whitespaces in BCP. Similarly, '.' was not handled correctly for host bundles.
There was bug in the getManifest() logic - it could throw NPE. We also changed the getManifest()
to always return the manifest of the host bundle. if host does not have a manifest, we return null.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42244 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5