Fix for 13963
Originally the secure admin-related bootstrapping was optional with the default=true.  During reviews no one came up with a use case in which we would NOT want to bootstrap, so this change removes that option and always bootstraps.

This class also used to use the SecureAdminClientManager. But because it is not run from inside the DAS this command should never need that, so I removed it.

Tests: QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42683 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5