Fix for 14352
Specifying a non-existent directory resulted in a puzzling error message about a supplemental command failing, with no information about why it had failed.

With this change the DAS "before" supplemental command that checks to make sure a deployed directory is the same on the DAS and on a remote target lets the main deploy command report if the file does not exist.  

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42752 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5