Fix for 12085
A few places in the server-side support for Java Web Start launches assumed that library JARs would have manifests and that the manifests would have main attributes (both are required by the JAR spec).  With these changes the code is more robust, checking for such conditions and logging them but not failing.

Tests: QL, deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42844 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5