issue 14161 - improve the performance of asadmin.
This changes the ASenvPropertyReader so that it only calculates the values for hostname and javaRoot if they are asked for.  It turns out that asadmin never actually asks for the values, so this saves on an expensive call to getCanonicalHostName.  Also, the SmartFile.sanitize method was rewritten to be more efficient. 

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42955 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5