Fix for issue 13882 and 13713
Two new weld-integration related fragments are introduced:
1. weld-integration-fragment to fix issue 13882 [1]
2. weld-integration-test-fragment to handle the use of weld internal packages by Weld arquillian tests (issue 13713).

The weld-integration-fragment is always bundled with the glassfish-jcdi package, as we want those packages to be available for Weld generated proxies at runtime as described at [1]. The weld-integration-test-fragment is not added to jcdi package as that would be used only the SQE team and other developers who want to run the arquillian tests against GlassFish.

- detailed description of fix for 13882 -
To fix issue 13882, the new implementation of ProxyServices SPI uses the thread context classloader (the application  classloader) as the classloader for loading Weld-generated bean proxies. The classloader that loaded the Bean must be used to load and define the bean proxy to handle Beans with package-private constructor as discussed in WELD-737.

Weld proxies today have references to some internal weld implementation classes such as javassist and org.jboss.weld.proxy.* packages. These classes are  temporarily re-exported through the weld-integration-fragment bundle so that when the bean proxies when loaded using the application classloader will have visibility to these internal implementation classes.

As a fix for WELD-737, Weld may use the Bean's classloader rather than asking the ProxyServices service implementation. Weld also plans to remove the dependencies of the bean proxy on internal implementation classes. When that happens we can remove the weld-integration-fragment workaround and the ProxyServices implementation.

Reviewed by: Sahoo(pom.xml and proxy classloading changes), Snjezana(packaging changes)
Tests run: QL, CDI developer tests, ensured that the distribution contains the weld-integration-fragment bundle




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@42968 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5