These changes make sure that the "enabled" attribute of the cluster-level application-ref element is correct during a rolling application upgrade.
If the target of enable or disable is an instance of a cluster, then the DAS now forwards the enable or disable request to all instances in that cluster so each can update its configuration information.  Only the specified target loads the app on enable or unloads it on disable.

Tests: deployment cluster devtests, rolling upgrade deployment devtests, QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43020 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5