Issue #14134: Allow JPA in Java SE mode from OSGi context. We now have introduced a new provider resolver which is part of a new module called osgi-jpa-extension.jar. This resolver is disabled by default, but can be enabled by configuring org.glassfish.osgjpa.extension.useHybridPersistenceProviderResolver=true in osgi config file. This provider allows PersistenceProviders to be discovered even though a provider may not be visible to caller's context class loader as long as the provider is installed as an OSGi bundle and has a META-INF/services/javax.persistence.spi.PersistenceProvider file.
With this change, user can now do Persistence.createEMF(). A test case to this effect is attached in JIRA.

This new resolver only comes into effect in OSGi environment, as it is set inside the bundle activator
of the new bundle. The reason we separated this into a separate bundle is because deployed OSGi bundles
which get started before autostart/ bundles may use Persistence.createEMF in their activator. If we had 
set the resolver in our existing autostart/osgi-jpa.jar, they would not have seen the effect. To ensure,
we set the resolver before any osgi-ed app is executed, we have a dummy Startup service in the new bundle as well.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43268 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5