Fix for issue 14842 and 11536
During APPLICATION_LOADED event, when we deploy beans to the Weld runtime, the TCL is set to the 
EARClassloader with only EJB JAR modules and application libraries. The default ResourcesLoader in 
Weld uses the TCL to load Beans and it failed to load Beans defined in web-modules of an EAR. 

Since there is no explicit indication of the module whose Beans are being processed, in 
BeanDeploymentArchiveImpl.getBeans, as a workaround we temporarily 
reset the TCL to the module classloader, so that all classloading for Beans in a module 
uses its module classloader. We finally reset TCL back to the original TCL after 
deployBeans is complete in WeldDeployer.event(). 

Tests run: PE QL, Weld devtests, testcases in 14842 and 11536
Reviewed by: Sahoo and Hong 



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43350 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5