Fix for 14938
The code which returns the URI path for launching a client using Java Web Start worked only if the app was deployed to the DAS, not if the app was deployed only to one or more instances.

This change returns the correct path regardless of where the app is deployed.

Tests:  QL, deployment devtests, the error scenario

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43402 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5