Various issues with the set and get commands:
issue 14889 - different output from get command - this fixes a corner case where an element
of the attribute path shows up twice in the name of the element. 

issue 14434 - the set command was not being replicated for names that start with "domain"

issue 14599 - the set command now support setting of leaf elements which it did not support before in v3.  This was supported in v2. 

issue 14921 - make the set/get commands support access to domain attributes, e.g., log-root, locale, etc.
These can now be set/get using domain.locale, domain.log-root and they are replicated to all instances. 

Passed QL. 

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43548 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5