Fix SSH code to handle spaces in paths.
13344 SSHLauncher.runCommand() is not space-in-path friendly

This changes adds methods to SSHLauncher to take the command to run
as an argv style List which it then quotes as needed to protect spaces
before it ships the command over SSH.

Also fix a couple other places in the code that needed to handle
spaces in paths better.

With these changes I was able to to run install-node, create-node-ssh,
create-instance, start-instance, stop-instance, delete-instance all
with a remote GlassFish installation that had a space in the path and
a nodedir with a  spacee in the path.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43627 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5