Fix for GLASSFISH-15230 continued
The earlier commit(43924) had issues with hybrid OSGi CDI applications.
The OSGi CDI portable extension is fired ProcessInjectionTarget events twice for an injection target. 
Once when Weld is asked to fireInjectionTargets for EE components and then when deployBeans are invoked. 

Before the fix in svn commit 43924 both the events resolved to the same FieldInjectionPoint (hashcode wise) 
and hence the serviceInjectionPoints HashSet and the extension added only Bean for that injectionpoint. 
However after that change, the hashCodes of the two FieldInjectionPoints are different and hence a 
duplication InjPoint and a Bean is added causing this issue.

The OSGi CDI extension now use a equals based Set, such as CopyOnWriteArraySet.

Tests: Ensured that OSGi CDI testcase passes, CDI devtests pass, QL on classic and web profiles pass




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@43952 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5