Fix for 15624
If an app contained a file named using a special character - one that would be encoded in URLs - then the payload implementation failed during synchronization as it tried to expand the payload uploaded to the instance into the instance's directories.  

This change repairs that problem.

Approved: Tom
Code review: Tom
Tests: error scenario in the issue, QL tests, deployment devtests with upload

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@44653 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5