Fix for 15653
CLIBootstrap generates the java command which launches the app client container.  It was incorrectly using the AS_JAVA or JAVA_HOME setting to locate the Java-provided extension and endorsed directories.  With these changes it instead (and correctly) uses the java.home system property which does point to the JRE home for locating the lib directory (which in in turn contains ext and endorsed).

Approved: Chris
Reviewed: Tom
Tests: QL, deployment and EJB devtests, SQE test which failed for Sonia (tested on my Mac and on her Solaris system)

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@44709 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5