Fix for 15716
appclient -help displays the usage output, not the much more complete help output.  This is because the code checks to make sure the user specifies a main class or a client JAR; if not, it automatically acts as if the user entered "-usage" instead.  But that logic failed to account for cases in which the user specified -help (or -usage for that matter).  And because the ACC checked for "-usage" before "-help" it would display the usage text instead.

These changes fix that.

Approved: Chris
Reviewed: Hong
Tests: updated appclient devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@44748 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5