Fix for 15963
Work done earlier to help optimize synchronization of uploaded archives to remote instances introduced a bug which broke the .reload feature.  Files uploaded during deployment are placed into a temporary directory under the domain's applications directory.  The code at fault incorrectly assumed that a file anywhere under the domain's applications directory was an uploaded file.  Deployment expands all archives and, except for the sync feature, the uploaded archive is no longer needed.  So the code moved what it thought was the uploaded file to a separate directory for use by synchronization.  But when a user touches the .reload file in the applications/(appName) directory for an already-deployed app, that directory is within the applications directory so this logic was moving the expanded directory to this separate directory, causing rather some havoc.

This change repairs the problem by not moving a file under the applications directory if it is itself a directory.  This is the lowest-risk fix at this point in 3.1.

Approved: Nazrul
Reviewed: Hong
Tests: QL, deployment devtests including a new test for the .reload feature

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@45068 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5