Fix for 6545
The FileArchive class implements the Archive interface for a directory tree (as opposed to a JAR for example).  The method which finds all files within a directory could fail with an NPE because file.listFiles can return null (if the file is not a directory or, in this case, if file permissions prevented GlassFish from getting the list of files).  

These changes detect a null return from file.listFiles and logs a WARNING-level message before returning, thus avoiding the NPE.

Tests: new unit test for inability to access directory; deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@45276 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5