Fix for 16261
The @PreDestroy method on an app client class (whether specified using annotations or in the descriptor) was not called.

The clean-up logic in the ACC was correctly written to invoke this but the injection manager in the clean-up objects was never set to the injection manager for the main class, so the clean-up logic could not invoke the pre-destroy method in the app client.

These changes store the injection manager with the clean-up information so the clean-up logic can invoke the method.

Tests: QL, deployment dev tests


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@45720 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5