Fix for 16195
The Java Web Start-aware app client container must make sure that certain classes which are present in both the Java SE installation and in a GlassFish-provided JAR are loaded from the GlassFish JAR.  This behavior was not working correctly because the Java Web Start-aware ACC used the MaskingClassLoader from the core/bootstrap module.  That implementation did not do what the ACC needs.

This change uses a fine-tuned masking class loader which works as the ACC needs in this scenario.

Tests: app client launches via Java Web Start of an app containing a web service method which throws a user exception (the user-provided scenario which triggered the error)

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@45806 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5