System-property elements are now handled completely.  There are 4 "levels" of system-property in domain.xml.  We parse from each level and then have them overlay each other with the correct priority which is
domain < cluster < config < server

Note that this is the stand-alone parser.  It is used by clients -- not the servers themselves.

Now you can do things like set a token in a jvm-option and have it resolve properly from, say, a system-property in the cluster element.  Then 
override in one particular server by defining it in the server element.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@46062 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5