File: Profiler.java, Line: 118, Type: WMI_WRONG_MAP_ITERATOR, Priority: Low
Method com.sun.enterprise.admin.launcher.Profiler.getPropertiesAsJvmOptions(Map) makes inefficient use of keySet iterator instead of entrySet iterator

This method accesses the value of a Map entry, using a key that was retrieved from a keySet iterator. It is more efficient to use an iterator on the entrySet of the map, to avoid the Map.get(key) lookup.




git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@46457 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5