Fix for Bug 12408108 - WINDOWS XP ONLY: SIGTEST DEPLOYMENT ERROR WHEN DEPLOYING SIGTEST EAR FILE
During a Windows deployment, a JAR in WEB-INF/lib was opened but never closed.  When the app was undeployed Windows could not remove the file.  

During a subsequent redeployment the left-over library JAR is incorrectly detected as a submodule archive because the ApplicationArchivist uses File directory scans rather than the FileArchive API.  

These changes (1) make a factory method in FileArchive for the StaleFileManager public, and (2) enhance ApplicationArchivist to use the StaleFileManager to filter what files are visible.  This correctly hides the left-over library JAR from being discovered and misinterpreted as a submodule archive.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@46541 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5