issue 13533 - make commands such as rotate-logs print an error message if they are run against an instance that has never been started. This effects rotate-logs, migrate-timers, generate-jvm-report, and freeze-transaction-service
A new option has been added to the @ExecuteOn annotation, ifNeverStarted, which takes 
a FailurePolicy.  The default is ignore, so commands that do set this continue with the 
previous behavior. This change also changes the interface to the GlassFish-private class, 
ClusterOperationUtil.  All uses of that class have been changed.

Passed QL on Windows.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@46565 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5