Fix for 16600
A recent change causes the first java invocation (of CLIBootstrap) to write the agent arguments to a temp file, rather than pass them on the generated command line, because of apparent limitations in the length of the -javaagent token in Java on AIX.  The earlier code escaped $ characters in file paths that were added to the command line (any env vars would already have been substituted by the shell invoking Java to run CLIBootstrap, so any remaining $ symbols are actually part of the path and need to be preserved).  Such escaping is no longer needed for agent arguments, now that they are written to a temp file, but the code continued to escape them, causing problems when those file paths were used by the ACC.

With this change CLIBootstrap no longer escapes $ symbols written to the agent arguments temp file.

Tests: QL, deployment devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@46780 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5