JIRA 16725
Problem:  If the server is started without the -javaagent JVM option then ProbeProviderClassFileTransformer.java gives up 
right away and monitoring will not work because the agent class can't be loaded.

Solution:
Instead of giving up -- try to attach to the agent.  Once that is done the class can be loaded and monitoring will work.





git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@47110 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5