Fix for 16719
When a new instance is created the domain.xml from the DAS was included in the bootstrapped files.  This seemed to confuse things if the new instance was started when the DAS was down, thus preventing a sync.

These changes cause the bootstrapping to no longer include domain.xml.

Tests: admin CLI cluster tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@47153 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5