Jira 16774
If you deleted or moved "tools.jar" from the lib directory of the JDK used to run GF -- fireworks ensued in the server log as monitoring tried to
use the Attach API (requires tools.jar).  Over and over and over and over it tried.

Fix: Check once and for all just in time.

Technique -- thankfully doing it via reflection was not necessary becasue we ALWAYS have the classes in tools.jar available at compile time.  What happened
before this fix is a ClassDefNotFound (sp?)  **error** (NOT Exception!) was thrown.  This happened simply trying to load the class that accesses the tools classes.
The solution iwas to create a wrapper class that is fully protected with catch(Throwable) -- and have it run the methods in the class that access the tools jar.  If
the latter can't be found then the wrapper class remembers and does not try again.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@47210 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5