JIRA 16287
This is where the process utilities I've added pay off.  No more fooling around.  To do a restart we now
wait for the OS itself to declare that the old server's process is really really dead and buried.
Before we did half checks like seeing if the admin port went deaf. Or that the standard streams closed. 
That turned out to be not good enough.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@47314 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5