Shoal integration for the bug fix below. This change was previously integrated into the 3.1.1 branch.
Fix for glassfish 16422.  This fixes bug that only occurs when one deletes and recreates a cluster with same name.
Workaround would be to restart DAS everytime a cluster is deleted and one plans on recreating the deleted cluster.

This is a low risk fix to ensure that distributed state cache is cleared of stale data when a cluster is deleted.
The DAS only leaves a cluster when the cluster is deleted (or DAS is shutting down.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@47962 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5