ntroduce parameter "gmsbroadcast" to asadmin create-cluster command. By default, it defaults to udpmulticast. If set to "tcp", then gmsbroadcast defaults to non-multicast mode using tcp to each member.
This is phase 1. This checkin actually sets GMS_DISCOVERY_URI_LIST with the DAS
ip address and port 9090 when non-multicast is default and GMS_DISCOVERY_URI_LIST is not
set or set to "generate".  Next phase will move generation of GMS_DISCOVERY_URI_LIST
to GMSAdapterImpl.initialize().  Will initialize based on 
ip addresses in nodes element of domain.xml.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@49324 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5