Trunk: 16034 forbid create-local-instance to create new instances in node-config Trunk: 7025067 (bugster) provide option to avoid changes to cluster topologies
Forward port of enhancement from 3.1.2.

This adds a "freeze" attribute to the <node> and <nodes> elements.
If the attribute is true on a <node> then instance creation on that
node is not allowed. If the attribute is true on <nodes> then new node
creation is not allowed.

Currently you must use the "set" command to set this attribute.
For example:

asadmin set nodes.freeze=true
asadmin set nodes.node.<node-name>.freeze=true



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@51545 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5