Added ability to provison GF from a zip distro. The url of distro can be specified using fighterfish.provisioner.url property. The location where GF to be installed can be specified using com.sun.aas.installRoot.
There is one issue with the way exam-rbc bundle is written. It Activator depends on slf4j, yet
it does not import it using regular import. It relies on DynamicImport:* to get this package. 
Moreover, this is the first bundle to be installed. slf4j packages are actually exported by
pax-logging-api bundle which is installed after this. So, there is a timing issue here. If
framework's start level reaches 2 (which will be the case when we bootstrap using GlassFish API)
before installing/starting rbc bundle, then it fails to activate.
We avoid this by explicitly installing bundles first and then starting them later in
installAndStartBundles(). 

Tested GF 3.1.2-SNAPSHOT & 4.0-SNAPSHOT - both web and full distro.
Does not work with GF 3.1.1 as it does not have osgi.properties file in config/.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@52132 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5