Modify the CLI framework so that when password parameters are passed between the client and the server, they are not prefixed with "AS_ADMIN_".  For compatibiliy, the server will still accept a password parameter that is prefixed in this way, however, the command model will use the declared name of the parameter. This change does require that the i18n key for the parameter use the declared name rather than the prefixed name.
The CLI client (asadmin) still prefixes the password with AS_ADMIN_ for use in a 
password file, so the syntax for that is unchanged. 

This change is needed to support other CLI clients that do not use AS_ADMIN_ as the 
prefix for passwords in the password file. 

Passed QL.



git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@53802 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5