Initial support for command-based authorization.
These changes allow command developers to add @AccessRequired annotations to their commands to indicate what authorization checks should be performed.  The command framework is modified to automatically perform the security check before invoking the execute() method of the command.

The command context is also expanded to allow the storage and retrieval of the Subject, which will result (once integrated) from authentication.

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@53896 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5