GLASSFISH-18850: Add hk2-locator metadata file to be compatible with new HK2.
The actual metadata location should be META-INF/hk2-locator/default and what I had was META-INF/hk2-locator.
Because of the way svn works, I need to commit after removing the old filer in order to create a directory in its place.
Adding the file in correct location this time.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@54830 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5