Fix for 18925
The doc instructs developers who want to customize the generated JNLP document to add <jnlp-doc>path-to-custom-JNLP</jnlp-doc> to their glassfish-application-client.xml descriptor.  The code works this way.  But the DTD comments and definition suggests that developers should instead use <jnlp-doc href="path-to-custom-JNLP"/>.

These changes remove the #REQUIRED setting from the jnlp-doc href attribute (since it is in fact not required) and makes it #IMPLIED instead.  They also enhance the code to look either for the text value or the href attribute.  Developers should not use both forms in a single glassfish-application-client.xml descriptor.

Tests: QL, deployment and ejb devtests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@55290 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5