Added a couple of new samples as described below:
webosgibridge: This is a web app that can be used to bootstrap an OSGi container inside 
a Servlet Container.

provisionerwebapp: This is a web app that can be used to bootstrap GlassFish
inside the OSGi framework provisioned using the former application.

Combining these two, what can be done is left to your imagination.


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@55423 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5