Revise the arguments passed to authorization preprocessors.
By passing the command object (as well as the command params) the preprocessor can introspect the class for annos, etc., if that makes sense.  Which is does for account-related processing.  Commands that refer to accounts that should exist have a parameter annotated with @AccountExists.  

The account-aware preprocessor (check-in coming shortly) looks for that anno and uses the annotated field's value for the account ID.  This lets the preprocessor avoid having to guess at what the account-related parameter is named.

Tests: QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@56680 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5