Fix for 19435.
These changes restore the feature by which remote admin access to the DAS is prohibited if secure admin is disabled.

The GenericAdminAuthenticator's various loginAsAdmin methods (and their definitions in the AdminAccessController interface) now consistently expose the following contract:  The returned Subject is always non-null.  If the incoming request is remote and secure admin is disabled then the methods throw the new RemoteAdminAccessException.  If the authentication fails due to a username/password problem then the methods throw a LoginException.

The changes to the RestAdapter and SessionsResource with ResourceUtil let them work with this new GAA.loginAsAdmin contract.

The AdminAdapter class is, in theory, obsolete, but because it's still in the repository I've changed it to keep it in-step with these other changes as well.

Tests: QL, multi-instance tests with instances on the same and different hosts from the DAS

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@57580 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5