Add the use of the new KernelIdentity.
The newly-added KernelIdentity represents an internal Subject that, typically, will be permitted
to perform all admin operations.  This is used from trusted system code that submits commands for
execution when there is not a Subject already available that represents an admin user.  These 
changes modify those places to pass the kernel identity's Subject to the command processing.

These changes also refactor some of the admin principals that convey non-username/password
identities.

There are also some changes to test utility or superclasses to use the Subject when submitting
commands from the tests themselves.

Tests: QL, admin devtests for both platforms


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@57847 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5