ORef: Effective Java, 2nd Edition, item #8 I tried and failed to fix these FindBugs issues.  The problem is that there are sub-classes that define equals(). The equals method is non-symetrical, so there may be problems using them in collections. I.e. PrincipalImpl pi = .... Group g = ...
pi.equals(g) ==> true
g.equals(pi) ==> false

The code that I'm rolling back really does something bad -- in the base class equals() it is looking for the 2
specific sub-classes.  Bad idea! Makes the code rather brittle.

The nexyt problem, which is too difficult to fix, is that the callers are MNOT USING the Principal interface.  Instead, they use Role, Group, PrincipalImpl.  So why even bother having an interface?  Anyways my earlier fix was to use Joshua Bloch's recommendation and use composition instead of inheritance.  This in turn demanded that the objects be used as interface references.  That is not the case -- there is more code working with the implementation class explicitly rather than the interface.

I'm just returning this whole mess to the way it was before.  I'll file an issue with the Security team for a proper solution.





git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@58216 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5