Fix QuickLook embedded test failure
The earlier check-in included changes which detect the embedded system administrator, but incorrectly skipped invoking a command's preAuthorization or getAccessChecks method.  Those methods can do useful work for the command before its execute method is run.  

This change makes sure that both of those methods are always invoked, even if the subject is the embedded system administrator.

Tests: QL embedded

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@59878 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5