Fix for 19763.
Requests that arrive from the console (as opposed to the CLI or via the management/domain path) have the SSL cert from the GlassFish keystore (alias s1as) attached. IIRC this is because the console code, which is running in the DAS itself, sends admin operations out as HTTP requests to the DAS - as does any admin client - to trigger admin operations in the DAS.

With secure admin enabled, the DAS admin listener wants (but does not require) client SSL certs. As an HTTP sender the DAS is set up to use SSL when it sends its own admin requests out, and although this is primarily intended for communicating securely with other instances in the domain this also happens when the DAS (that is, the console) is talking to its own admin listener with secure admin enabled. So, with secure admin enabled all admin requests from the console have the DAS's SSL cert attached.

When the DAS apart from the console sends requests to other servers in the domain it adds the unique domain identifier to the request, and the receiving server checks to make sure that the domain ID in the request matches the domain ID in its configuration. If the IDs match then the server is confident that the request came from the DAS (because the request also had the DAS's SSL cert) and so identifies the admin client as an administrator.

The redesigned authentication implementation has two LoginModules which cooperate to establish the client's identity. The first handles the various non-username/password techniques for authenticating to a server: local password, admin or ReST tokens, and SSL certs plus the domain ID. The second LoginModule handles username/password log-ins but is not run if the first LoginModule recognizes the client.

This issue occurred because the requests from the console have the DAS SSL cert but do not have (and should not have) the unique domain ID. The first LoginModule detects the SSL cert but, because the unique domain ID was absent, did not identify the client as an administrator. Because the first LoginModule reported that it recognized the client, the second LoginModule which would have processed the username and password was never run. The result: with secure admin turned on the console requests were never treated as coming from a valid administrator and so were never accepted. This was reflected in the console as a rejected log-in.

Some recent changes to how authorization works exposed this problem.

With these changes the admin LoginModule no longer attaches a Principal for the DAS's SSL cert if there is also a username/password with the request.  This causes the second login module to be run which processes the username and password.

Tests: QL, admin dev tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@60125 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5