Open PasswordAdapter lazily
As part of abstracting the PasswordAliasStore the JCEKS implementation initialized the underlying PasswordAdapter immediately.  This brought some work into server start-up when it was not necessarily so earlier.

This change defers the initialization of the PasswordAdapter until the first actual use to avoid adding work to server start-up.

Tests: QL

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@60424 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5