GLASSFISH-19216: NPE(Specified service reference cannot be null) during server shutdown is preventing clean up to complete
This issue does not occur in 3.1.2. This is a regression caused by GLASSFISH-18188 which is fixed in GlassFish 4.0 only. 
As part of that bug fix, we unregister GlassFish from service registry as the first step in shutdown path. 
We are doing so, because GlassFish's state becomes Stopping and as such it is useless as one can't call methods like getService(). 
So, I think unregistering early makes sense. 

AbstractOSGiDeployer's undeploy method is getting called during shutdown and 
there it is trying to get a reference to GlassFish service, so it is getting null.
I have switched to using Globals API.

Tested with both GF 3.1.2.2 and GF 4.0


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@60807 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5