Additional fix for GLASSFISH-20160 - asadmin get command referring to a default configured item fails
An earlier check-in restored a long-ago removed bit of the domain.xml template for the default security audit-module, but that confused the upgrade logic that had been changed when the audit-module had been removed earlier.

This change fixes the upgrade problem.

Approved by Tom
Reviewed by Tom
Passed QL, the upgrade GlassFish admin devtest

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@61214 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5