Fix for GLASSFISH-20613
The problem was not that the resource names needed to be encoded before creating the URI (which is what the earlier fix for 20510 did) but that the resource name from the EJB monitoring code needed to provide a URI-friendly item name.

In a comment on 20510 Byron provided the fix for the monitoring code.  This check-in includes that change and a change in CommandSecurityChecker to remove the encoding logic that was inserted earlier.

Passed QL, the ejb dev test which failed in 20510


git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@62176 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5