Fix for GLASSFISH-20671
The enable- and disable-secure-admin commands now automatically change the JMX connector security settings (as well as the DAS admin listener settings which they already did).  

This change updates a config listener which used to assume that an updated Ssl instance was always a child of a Protocol.  An Ssl can also be a child of a JMX connector.

Tests: RQ, manual tests

git-svn-id: svn+ssh://svn.java.net/glassfish~svn/trunk@62418 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5